Created functions for interfacing with TeensyCANBase without using classes
Only the constructor of the class was actually being used, so removing it
may increase the understandability of the class.
